refactor generated mock data
