+++ js/mock-data/generate-ad.js
@@ -1,8 +1,10 @@
 import { generateAuthor } from './generate-author.js';
+import { generateLocation } from './generate-location.js';
 import { generateOffer } from './generate-offer.js';
 
 const generateAd = () => ({
   author: generateAuthor(),
+  location: generateLocation(),
   offer: generateOffer(),
 });
